// src/app/books/interfaces/book.interface.ts
// Generated by https://quicktype.io

export interface Book {
  id:          number;
  title:       string;
  author:      string;
  image:       string;
  genre:       Genre;
  isbn:        string;
  published:   string;
  publisher:   string;
  description: string;
}

export enum Genre {
  ut  = 'Ut',
  nam = 'Nam',
  vol = 'Voluptatem',
  und = 'Unde',
  ape = 'Aperiam',
  in  = 'In'
}